+ add debug logging of rabbitmq messages + make commits field optional + omit commits and files from ReleaseMessagePayload if empty + remove unnecessary tools (pybomb)
