--- tools/pybomb/.gitignore
@@ -1,1 +1,0 @@
-venv
